requirements.txt
asgiref==3.4.1
Django==3.2.8
django-crispy-forms==1.13.0
django-recaptcha==2.0.6
Pillow==8.4.0
psycopg2==2.9.1
pytz==2021.3
sqlparse==0.4.2